libobs: More fullscreen triangles instead of quads
Replace fullscreen quads with triangles when outputting RGBA video, or
resizing the final canvas with the downscale setting. Quads are still
used for scale applied directly to sources, and scale filters because
it's not as easy to clip the triangle back down to a quad.

Speed-up is a byproduct of better rasterization ordering.

Intel HD Graphics 530, Intel GPA, SetStablePowerState

1920x1080, RGBA output: ??? us -> ??? us
1920x1080 -> 1280x720, RGBA output, Bilinear: ??? us -> ??? us
1920x1080 -> 1280x720, RGBA output, Area: ??? us -> ??? us
1920x1080 -> 1280x720, RGBA output, Bicubic: ??? us -> ??? us
1920x1080 -> 1280x720, RGBA output, Lanczos: ??? us -> ??? us
3840x2160 -> 1280x720, RGBA output, Bilinear Lowres: ??? us -> ??? us
1920x1080 -> 1280x720, NV12 output, Bilinear: ??? us -> ??? us
1920x1080 -> 1280x720, NV12 output, Area: ??? us -> ??? us
1920x1080 -> 1280x720, NV12 output, Bicubic: ??? us -> ??? us
1920x1080 -> 1280x720, NV12 output, Lanczos: ??? us -> ??? us
3840x2160 -> 1280x720, NV12 output, Bilinear Lowres: ??? us -> ??? us